Updated calling convention and function example
New calling convention. Arguments are passed in registers in increasing
order.
Ie. The first argument is D0, second argument is D1. D0 is used as the
return register.
This allows returned values to be immediately passed as arguments to
later functions.

Functions must save and restore registers that were not used as
arguments.